LayoutTests/fast/xmlhttprequest/xmlhttprequest-properties-prototype-expected.txt

Checks that XMLHttpRequest's properties are on the prototype and have the right properties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS xhr.__proto__ is XMLHttpRequest.prototype

* readyState
PASS xhr.hasOwnProperty(propertyName) is false
PASS xhr.__proto__.hasOwnProperty(propertyName) is true
PASS properties.configurable is true
PASS properties.enumerable is true

* response
FAIL xhr.hasOwnProperty(propertyName) should be false. Was true.
FAIL xhr.__proto__.hasOwnProperty(propertyName) should be true. Was false.
FAIL properties.configurable should be true. Threw exception TypeError: undefined is not an object (evaluating 'properties.configurable')
FAIL properties.enumerable should be true. Threw exception TypeError: undefined is not an object (evaluating 'properties.enumerable')

* responseText
FAIL xhr.hasOwnProperty(propertyName) should be false. Was true.
FAIL xhr.__proto__.hasOwnProperty(propertyName) should be true. Was false.
FAIL properties.configurable should be true. Threw exception TypeError: undefined is not an object (evaluating 'properties.configurable')
FAIL properties.enumerable should be true. Threw exception TypeError: undefined is not an object (evaluating 'properties.enumerable')

* responseType
PASS xhr.hasOwnProperty(propertyName) is false
PASS xhr.__proto__.hasOwnProperty(propertyName) is true
PASS properties.configurable is true
PASS properties.enumerable is true

* responseURL
PASS xhr.hasOwnProperty(propertyName) is false
PASS xhr.__proto__.hasOwnProperty(propertyName) is true
PASS properties.configurable is true
PASS properties.enumerable is true

* responseXML
PASS xhr.hasOwnProperty(propertyName) is false
PASS xhr.__proto__.hasOwnProperty(propertyName) is true
PASS properties.configurable is true
PASS properties.enumerable is true

* statusText
PASS xhr.hasOwnProperty(propertyName) is false
PASS xhr.__proto__.hasOwnProperty(propertyName) is true
PASS properties.configurable is true
PASS properties.enumerable is true

* timeout
PASS xhr.hasOwnProperty(propertyName) is false
PASS xhr.__proto__.hasOwnProperty(propertyName) is true
PASS properties.configurable is true
PASS properties.enumerable is true

* upload
PASS xhr.hasOwnProperty(propertyName) is false
PASS xhr.__proto__.hasOwnProperty(propertyName) is true
PASS properties.configurable is true
PASS properties.enumerable is true

* withCredentials
PASS xhr.hasOwnProperty(propertyName) is false
PASS xhr.__proto__.hasOwnProperty(propertyName) is true
PASS properties.configurable is true
PASS properties.enumerable is true

PASS successfullyParsed is true

TEST COMPLETE
